Install/use scripts from inside the app rather than outside
* text-based scripts (such as shell scripts) added to /assets/ (rather than
  independently installed to /system/bin or /system/xbin)
* on EVERY run, scripts are copied to local app cacheDir & permissions set
  to 700
* All the fragments are also updated to point to the local script in cachedir

Bonuses:

* verify that /sdcard/files/ exists and create it if not
* minor fixes & cleanup

Note: Scripts are re-installed at each run in case the user has cleared app
data (ie, cache) or in case this a new version of the app (which may have
newer versions of the scripts to install).

Portions of this patch (namely, the scripts in /assets/) are moved from
the kali-nethunter installer repository at:
   https://github.com/offensive-security/kali-nethunter/
Any modifications from me to those scripts are licensed under the same terms as
the original files.