Some changes to network list.
Moved to from `netcfg` to `ip -o addr show` since the command is not found in Marshmallow.

`ip` command is giving more detailed information.

Moved Network and Hid interfaces to a ListView (added item layout).

Now, with a long press in the listItem, the interface address is copied to the clipboard so we can paste arround!!