Check custom commands for terminal permission
Default error was to say that Nethunter Temrinal wasn’t installed even if problem was permissions related.  Now we tell the user that they are missing permissions instead or missing terminal.

This is more of a problem because of recent signing of Terminal and permissions being unchecked.  This should hopefully get rid of false errors.

Tested by unchecking permissions in settings > apps and checking permissions again.

Signed-off-by: binkybear <binkybear@nethunter.com>