Fix bootkali_init su call for Magisk
If rooting device with Magisk, line #21 fails - causing the Terminal to hang. This is due to su waiting for input. The common symptom of this is seen when launching the Nethunter Kali Terminal - it will drop into normal Android Terminal until you type 'exit'. This fix, the su -c command, will work with both SuperSu and Magisk